services/translator/pyproject.toml
[project]
name = "real-time-translation-translator"
version = "0.1.0"
description = "Translation microservice using Gemini/OpenAI"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.58.0",
    "httpx>=0.28.1",
    "langchain>=1.2.6",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.7",
    "pydantic>=2.11.0",
    "uvicorn>=0.40.0",
]

[project.scripts]
real-time-translation-translate = "translator_service.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/translator_service"]
